refactor global domain report by creating more generic AdminFacetedReport
